// clients/js-codama/src/generated/types/key.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from '@solana/kit';

export enum Key {
  Uninitialized,
  EditionV1,
  MasterEditionV1,
  ReservationListV1,
  MetadataV1,
  ReservationListV2,
  MasterEditionV2,
  EditionMarker,
  UseAuthorityRecord,
  CollectionAuthorityRecord,
  TokenOwnedEscrow,
  TokenRecord,
  MetadataDelegate,
  EditionMarkerV2,
  HolderDelegate,
}

export type KeyArgs = Key;

export function getKeyEncoder(): FixedSizeEncoder<KeyArgs> {
  return getEnumEncoder(Key);
}

export function getKeyDecoder(): FixedSizeDecoder<Key> {
  return getEnumDecoder(Key);
}

export function getKeyCodec(): FixedSizeCodec<KeyArgs, Key> {
  return combineCodec(getKeyEncoder(), getKeyDecoder());
}